% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2_scales_continuous.R
\name{ggplot2-scales-continuous}
\alias{ggplot2-scales-continuous}
\alias{scale_colour_random_c}
\alias{scale_color_random_c}
\alias{scale_fill_random_c}
\title{Random Continuous scales to use for ggplot2}
\usage{
scale_colour_random_c(chaos = FALSE, ...)

scale_color_random_c(chaos = FALSE, ...)

scale_fill_random_c(chaos = FALSE, ...)
}
\arguments{
\item{chaos}{Logical, if set to \code{TRUE} enables FULL CHAOS MODE. Seeding will
not work affect \code{theme_random()}. Defaults to \code{FALSE}.}

\item{...}{Arguments to pass on to \code{ggplot2::scale_colour_gradientn()} or
\code{ggplot2::scale_fill_gradientn()}}
}
\value{
A \code{ScaleContinuous} object that can be added to a \code{ggplot} object
}
\description{
These functions provide the option to use random continuous palettes along
with the \code{ggplot2} package. It goes without saying that it requires \code{ggplot2}
to work.
}
\details{
Available package/palette combinations are available in the data.frame
\code{\link[paletteer]{palettes_c_names}}.
}
\examples{

if (require("ggplot2")) {
  ggplot(mtcars, aes(mpg, disp, color = drat)) +
    geom_point() +
    scale_colour_random_c()

  set.seed(1234)
  ggplot(mtcars, aes(mpg, disp, color = drat)) +
    geom_point() +
    scale_colour_random_c(chaos = TRUE)
}
}
