% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_random.R
\name{theme_random}
\alias{theme_random}
\title{Random Theme for ggplot2}
\usage{
theme_random(chaos = FALSE)
}
\arguments{
\item{chaos}{Logical, if set to \code{TRUE} enables FULL CHAOS MODE. Seeding will
not work affect \code{theme_random()}. Defaults to \code{FALSE}.}
}
\description{
Random Theme for ggplot2
}
\examples{
library(ggplot2)

ggplot(mtcars, aes(mpg, disp)) +
  geom_point() +
  theme_random()

set.seed(1234)
ggplot(mtcars, aes(mpg, disp)) +
  geom_point() +
  theme_random(chaos = TRUE)
}
